feat: Add import statements for ES6 basic constants module
This commit adds import statements for the `taskFirst` and `taskNext` functions from the newly added `0-constants.js` module. These functions are then used to log a message to the console.

The import statements allow for the usage of the `taskFirst` and `taskNext` functions in other parts of the codebase, enhancing code reusability and maintainability.