feat: Add default parameters to getSumOfHoods function
The code changes in this commit add default parameters to the `getSumOfHoods` function in the `3-default-parameter.js` file. The `initialNumber` parameter remains required, while the `expansion1989` and `expansion2019` parameters have default values of 89 and 19 respectively. This allows for more flexibility when calling the function, as the default values will be used if the corresponding arguments are not provided.

This enhancement improves the usability and readability of the `getSumOfHoods` function, making it easier to use and understand.