feat: Add asyncUploadUser function
This commit adds the `asyncUploadUser` function to the `100-await.js` file. The function uses `Promise.all` to await the results of `uploadPhoto` and `createUser` functions. It returns an object with the `photo` and `user` values. If there is an error, it returns an object with `null` values for `photo` and `user`.