feat: Add Currency class
This commit adds the `Currency` class to the `0x02-ES6_classes/3-currency.js` file. The class has a constructor that takes in a `code` and `name` parameter and assigns them to the `_code` and `_name` properties respectively. It also includes getter and setter methods for `code` and `name`, as well as a `displayFullCurrency` method that returns a formatted string with the currency name and code.