feat: Add Airport class
This commit adds the `Airport` class to the `0x02-ES6_classes/7-airport.js` file. The class has a constructor that takes in a `name` and `code` parameter and assigns them to the `_name` and `_code` properties respectively. It includes a `toString` method that returns a formatted string representation of the airport. This class allows for the creation of airport objects with specific names and codes.