feat: Add HolbertonClass with Symbol.toPrimitive method
This commit adds the `HolbertonClass` class to the `0x02-ES6_classes/8-hbtn_class.js` file. The class has a constructor that takes in `size` and `location` parameters and assigns them to the `_size` and `_location` properties respectively. It also includes a `Symbol.toPrimitive` method that returns the `_location` property when the hint is 'string' and the `_size` property when the hint is 'number'. This class allows for the creation of `HolbertonClass` objects with specific size and location properties.