feat: Add HoistingClass and StudentHolberton classes
This commit adds the `HoistingClass` and `StudentHolberton` classes to the `0x02-ES6_classes/9-hoisting.js` file. The `HoistingClass` class has a constructor that takes in `year` and `location` parameters and assigns them to the `_year` and `_location` properties respectively. It includes getter methods for `year` and `location`. The `StudentHolberton` class has a constructor that takes in `firstName`, `lastName`, and `holbertonClass` parameters and assigns them to the `_firstName`, `_lastName`, and `_holbertonClass` properties respectively. It includes getter methods for `fullName`, `holbertonClass`, and `fullStudentDescription`. These classes allow for the creation of HolbertonClass and StudentHolberton objects with specific properties.