feat: Add Car class with cloneCar method
This commit adds the `Car` class to the `0x02-ES6_classes/10-car.js` file. The class has a constructor that takes in `brand`, `motor`, and `color` parameters and assigns them to the `_brand`, `_motor`, and `_color` properties respectively. It includes a `cloneCar` method that creates a new instance of the `Car` class with the same brand, motor, and color. This class allows for the creation of car objects with the ability to clone themselves.